+++ packages/server/src/model/part-of-speech/mut.ts
@@ -1,6 +1,6 @@
 import {UserInputError} from 'apollo-server';
 
-import {DataAccessor, DataReader} from '../../database';
+import {DataReader} from '../../database';
 import {
   PartOfSpeechId,
   NewPartOfSpeechInput,
@@ -10,6 +10,7 @@ import {
 import {Language} from '../language';
 import {Definition} from '../definition';
 import {SearchIndexMut} from '../search-index';
+import {MutContext} from '../types';
 
 import {PartOfSpeech} from './model';
 import {PartOfSpeechRow} from './types';
@@ -17,17 +18,18 @@ import {validateName} from './validators';
 
 const PartOfSpeechMut = {
   async insert(
-    db: DataAccessor,
+    context: MutContext,
     data: NewPartOfSpeechInput
   ): Promise<PartOfSpeechRow> {
     const {languageId} = data;
     let {name} = data;
 
-    const language = await Language.byIdRequired(db, languageId);
+    const language = await Language.byIdRequired(context.db, languageId);
 
-    name = validateName(db, null, language.id, name);
+    name = validateName(context.db, null, language.id, name);
 
-    return db.transact(db => {
+    return MutContext.transact(context, context => {
+      const {db, events} = context;
       const now = Date.now();
       const {insertId} = db.exec<PartOfSpeechId>`
         insert into parts_of_speech (
@@ -41,28 +43,36 @@ const PartOfSpeechMut = {
 
       SearchIndexMut.insertPartOfSpeech(db, insertId, name);
 
+      events.emit({
+        type: 'partOfSpeech',
+        action: 'create',
+        id: insertId,
+        languageId: language.id,
+      });
+
       return PartOfSpeech.byIdRequired(db, insertId);
     });
   },
 
   async update(
-    db: DataAccessor,
+    context: MutContext,
     id: PartOfSpeechId,
     data: EditPartOfSpeechInput
   ): Promise<PartOfSpeechRow> {
     const {name} = data;
 
-    const partOfSpeech = await PartOfSpeech.byIdRequired(db, id);
+    const partOfSpeech = await PartOfSpeech.byIdRequired(context.db, id);
 
     if (name != null) {
       const newName = validateName(
-        db,
+        context.db,
         partOfSpeech.id,
         partOfSpeech.language_id,
         name
       );
 
-      await db.transact(db => {
+      await MutContext.transact(context, context => {
+        const {db, events} = context;
         db.exec`
           update parts_of_speech
           set
@@ -73,26 +83,48 @@ const PartOfSpeechMut = {
 
         SearchIndexMut.updatePartOfSpeech(db, partOfSpeech.id, newName);
 
+        events.emit({
+          type: 'partOfSpeech',
+          action: 'update',
+          id: partOfSpeech.id,
+          languageId: partOfSpeech.language_id,
+        });
+
         db.clearCache(PartOfSpeech.byIdKey, partOfSpeech.id);
       });
     }
 
-    return PartOfSpeech.byIdRequired(db, partOfSpeech.id);
+    return PartOfSpeech.byIdRequired(context.db, partOfSpeech.id);
   },
 
-  delete(db: DataAccessor, id: PartOfSpeechId): Promise<boolean> {
-    this.ensureUnused(db, id);
+  async delete(context: MutContext, id: PartOfSpeechId): Promise<boolean> {
+    const {db} = context;
+    const partOfSpeech = await PartOfSpeech.byId(db, id);
+    if (!partOfSpeech) {
+      return false;
+    }
 
-    return db.transact(db => {
-      const {affectedRows} = db.exec`
+    this.ensureUnused(db, partOfSpeech.id);
+
+    await MutContext.transact(context, context => {
+      const {db, events} = context;
+      db.exec`
         delete from parts_of_speech
-        where id = ${id}
+        where id = ${partOfSpeech.id}
       `;
 
-      SearchIndexMut.deletePartOfSpeech(db, id);
+      SearchIndexMut.deletePartOfSpeech(db, partOfSpeech.id);
 
-      return affectedRows > 0;
+      events.emit({
+        type: 'partOfSpeech',
+        action: 'delete',
+        id: partOfSpeech.id,
+        languageId: partOfSpeech.language_id,
+      });
     });
+
+    db.clearCache(PartOfSpeech.byIdKey, partOfSpeech.id);
+    return true;
   },
 
   ensureUnused(db: DataReader, id: PartOfSpeechId) {
